--- a/openquake/hazardlib/tom.py
+++ b/openquake/hazardlib/tom.py
@@ -242,13 +242,15 @@ def get_probability_no_exceedance(ctx, poes, probs_or_tom):
         list of N probability mass functions otherwise
     """
     pnes = numpy.zeros_like(poes)
-    if hasattr(probs_or_tom, 'get_probability_no_exceedance'):  # poissonian
+    if isinstance(probs_or_tom, FatedTOM):
+        for i, rate in enumerate(ctx.occurrence_rate):
+            pnes[i] = probs_or_tom.get_probability_no_exceedance(rate, poes[i])
+    elif isinstance(probs_or_tom, PoissonTOM):
         calc_pnes(ctx.occurrence_rate, poes, probs_or_tom.time_span, pnes)
-        return pnes
     else:  # nonpoissonian
         for i, probs_occur in enumerate(probs_or_tom):
             pnes[i] = get_probability_no_exceedance_np(probs_occur, poes[i])
-        return pnes
+    return pnes
 
 
 def get_probability_no_exceedance_np(probs_occur, poes):