Continuation of https://github.com/gem/oq-engine/pull/7701. This time it works. Here is the difference for a calculation on my workstation: computing pnes  447.6s -> 108.9s (4x speedup). Here is the difference for the SAM model on the spot machine (2.5x speedup):
```
# before
| calc_44964, maxmem=53.2 GB | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 246_537  | 513.2     | 244       |
| computing pnes             | 129_942  | 0.0       | 1_466_402 |
| make_contexts              | 46_491   | 479.8     | 97_514    |
# after
| calc_44966, maxmem=53.1 GB | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 170_730  | 548.1     | 244       |
| computing pnes             | 52_955   | 0.0       | 1_466_402 |
| make_contexts              | 47_037   | 508.2     | 97_514    |
```